Remove commented lines and make few cosmetics changes. Done to respect review of mathildemerle and fcollot on PR #370

+++ src/layers/legacy/medCoreLegacy/gui/lookUpTables/medLUTToXMLWriter.cpp
@@ -69,7 +69,7 @@ void medLUTToXMLWriterPrivate::writeNode(const medClutEditorVertex & vertex)
 {
     QString node;
     //position
-    dtkDebug()<< "vertex: "<< reinterpret_cast<long long>(&vertex);
+    dtkDebug()<< "vertex: "<< reinterpret_cast<uintptr_t>(&vertex);
     dtkDebug() << "node:" << vertex.pos();
     dtkDebug() << "color" <<vertex.color();
     node.sprintf("%.4f;%.4f;%d;%d;%d;%d",vertex.pos().x(),
